Revert "update notification flow"
