--- src/app/client/src/app/modules/notification/services/notification-old/notification-service-impl.spec.data.ts
@@ -1,60 +1,0 @@
-export const notificationData = {
-  data: {
-    actionData: {
-      actionType: 'contentURL',
-      contentURL: 'https://url/resources/play/content/do_00000',
-      description: 'Notfication description 1',
-      openPlayer: true,
-      title: 'InApp Notification 1',
-      thumbnail: ''
-    },
-    type: 1
-  },
-  id: 'id1',
-  category: 'Notification',
-  priority: 1,
-  userId: 'user-id',
-  createdOn: 1605589382634,
-  status: 'unread'
-};
-
-export const notificationList = [
-  {
-    data: {
-      actionData: {
-        actionType: 'contentURL',
-        contentURL: 'https://url/resources/play/content/do_00000',
-        description: 'Notfication description 1',
-        openPlayer: true,
-        title: 'InApp Notification 1',
-        thumbnail: ''
-      },
-      type: 1
-    },
-    id: 'id1',
-    category: 'Notification',
-    priority: 1,
-    userId: 'user-id',
-    createdOn: 1605589382634,
-    status: 'unread'
-  },
-  {
-    data: {
-      actionData: {
-        actionType: 'contentURL',
-        contentURL: 'https://url/resources/play/content/do_00000',
-        description: 'Notfication description 2',
-        openPlayer: true,
-        title: 'InApp Notification 2',
-        thumbnail: ''
-      },
-      type: 1
-    },
-    id: 'id2',
-    category: 'Notification',
-    priority: 1,
-    userId: 'user-id',
-    createdOn: 1605589382634,
-    status: 'read'
-  }
-];
